big 2021 update
finally getting stuff working again.

switched to go modules for package management (no more `vendor`
directory).

updated GORM and fixed some primary/foreign key datatype issues that
came up as a result.

+++ api.go
@@ -47,7 +47,7 @@ func (s Server) PlayHandler(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	songID, _ := strconv.Atoi(vars["song"])
 
-	play := Play{SongID: uint(songID)}
+	play := Play{SongID: songID}
 	s.DB.Create(&play)
 	fmt.Fprintf(w, "ok")
 }
